refactor renderer entry exports
