+++ src/renderers/controls/OneOfRenderer.vue
@@ -100,11 +100,8 @@
 import {
   ControlElement,
   createCombinatorRenderInfos,
-  JsonFormsRendererRegistryEntry,
-  rankWith,
   createDefaultValue,
   CombinatorSubSchemaRenderInfo,
-  isOneOfControl,
 } from '@jsonforms/core';
 import {
   DispatchRenderer,
@@ -135,7 +132,7 @@ import CombinatorProperties from '../components/CombinatorProperties.vue';
 import { default as CzFieldset } from '../controls/components/cz.fieldset.vue';
 import { default as ControlWrapper } from './ControlWrapper.vue';
 
-const controlRenderer = defineComponent({
+export default defineComponent({
   name: 'any-of-renderer',
   components: {
     DispatchRenderer,
@@ -281,11 +278,4 @@ const controlRenderer = defineComponent({
     },
   },
 });
-
-export default controlRenderer;
-
-export const entry: JsonFormsRendererRegistryEntry = {
-  renderer: controlRenderer,
-  tester: rankWith(3, isOneOfControl),
-};
 </script>
