+++ src/renderers/layouts/ObjectLayoutRenderer.vue
@@ -19,14 +19,7 @@
 </template>
 
 <script lang="ts">
-import {
-  JsonFormsRendererRegistryEntry,
-  Layout,
-  rankWith,
-  and,
-  isLayout,
-  uiTypeIs,
-} from '@jsonforms/core';
+import { Layout } from '@jsonforms/core';
 import { defineComponent } from 'vue';
 import {
   DispatchRenderer,
@@ -56,9 +49,4 @@ const layoutRenderer = defineComponent({
 });
 
 export default layoutRenderer;
-
-export const entry: JsonFormsRendererRegistryEntry = {
-  renderer: layoutRenderer,
-  tester: rankWith(3, and(isLayout, uiTypeIs('Object'))),
-};
 </script>
